\name{ClimMob-package}
\alias{ClimMob-package}
\alias{ClimMob}
\docType{package}
\title{
  ClimMob: Crowdsourcing climate-smart agriculture
}
\description{
  Data analysis for crowdsourcing experiments to test climate-smart agricultural practices
}
\details{
  \tabular{ll}{
    Package: \tab ClimMob\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    Date: \tab 2013-06-25\cr
    License: \tab GPL-3\cr
    LazyLoad: \tab yes\cr
  }
  
}
\author{
  Jacob van Etten
  
  Maintainer: Jacob van Etten <jacobvanetten@yahoo.com>
}

\keyword{ package }
